+++ example/gender.dart
@@ -1,0 +1,2 @@
+/// Example enum branched on by the `_select` messages.
+enum Gender { male, female, other }