+++ example/json/json_example.dart
@@ -3,6 +3,8 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'package:i69n/i69n.dart' as i69n;
 
+import '../gender.dart';
+
 import 'exampleMessages.i69n.dart';
 import 'exampleMessages_cs.i69n.dart' deferred as cs;
 
@@ -17,6 +19,10 @@ void main() async {
   print(m.apples.count(1));
   print(m.apples.count(2));
   print(m.apples.count(5));
+  print(m.person.sees(Gender.female));
+  print(m.person.owns(Gender.male, 3));
+  print(m.person.status(true));
+  print(m.person.status(false));
 
   print('Asynchronous load of Czech messages:');
   await cs.loadLibrary();
@@ -29,6 +35,10 @@ void main() async {
   print(m.apples.count(1));
   print(m.apples.count(2));
   print(m.apples.count(5));
+  print(m.person.sees(Gender.female));
+  print(m.person.owns(Gender.male, 3));
+  print(m.person.status(true));
+  print(m.person.status(false));
 
   print('Access messages at runtime, with plain old string keys');
   print('Static:  ${m.generic.ok}');