+++ test/testMessages.i69n.dart
@@ -2,7 +2,7 @@
 // GENERATED FILE, do not edit!
 // dart format off
 import 'package:i69n/i69n.dart' as i69n;
-import 'dart:io';
+import 'mock/gender.dart';
 
 String get _languageCode => 'sk';
 String get _localeName => 'en';
@@ -11,14 +11,14 @@ String _plural(int count, {String? zero, String? one, String? two, String? few, 
     i69n.plural(count, _languageCode, zero: zero, one: one, two: two, few: few, many: many, other: other);
 String _ordinal(int count, {String? zero, String? one, String? two, String? few, String? many, String? other}) =>
     i69n.ordinal(count, _languageCode, zero: zero, one: one, two: two, few: few, many: many, other: other);
-String _cardinal(int count, {String? zero, String? one, String? two, String? few, String? many, String? other}) =>
-    i69n.cardinal(count, _languageCode, zero: zero, one: one, two: two, few: few, many: many, other: other);
+String _select(Object? value, Map<String, String> cases) => i69n.select(value, cases);
 
 class TestMessages implements i69n.I69nMessageBundle {
   const TestMessages();
   GenericTestMessages get generic => GenericTestMessages(this);
   InvoiceTestMessages get invoice => InvoiceTestMessages(this);
   ApplesTestMessages get apples => ApplesTestMessages(this);
+  PersonTestMessages get person => PersonTestMessages(this);
   FriendsTestMessages get friends => FriendsTestMessages(this);
   Object operator [](String key) {
     var index = key.indexOf('.');
@@ -32,6 +32,8 @@ class TestMessages implements i69n.I69nMessageBundle {
         return invoice;
       case 'apples':
         return apples;
+      case 'person':
+        return person;
       case 'friends':
         return friends;
       default:
@@ -80,7 +82,7 @@ class InvoiceTestMessages implements i69n.I69nMessageBundle {
     if (index > 0) {
       return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
-    throw Exception('[] operator is disabled in en.invoice, see _i69n: nomap flag.');
+    throw Exception('[] operator is disabled in sk.invoice, see _i69n: nomap flag.');
   }
 }
 
@@ -112,6 +114,34 @@ class ApplesTestMessages implements i69n.I69nMessageBundle {
         return quotes;
       case 'quotes2':
         return quotes2;
+      default:
+        return key;
+    }
+  }
+}
+
+class PersonTestMessages implements i69n.I69nMessageBundle {
+  final TestMessages _parent;
+  const PersonTestMessages(this._parent);
+  String sees(Gender gender) => "I see ${_select(gender, {'male': 'Him', 'female': 'Her', 'other': 'Them'})}";
+  String owns(Gender gender, int cnt) => "${_select(gender, {
+            'male': 'his',
+            'female': 'her',
+            'other': 'their'
+          })} ${_plural(cnt, one: '$cnt apple', many: '$cnt apples')}";
+  String status(bool online) => "${_select(online, {'true': 'Online', 'false': 'Offline'})}";
+  Object operator [](String key) {
+    var index = key.indexOf('.');
+    if (index > 0) {
+      return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
+    }
+    switch (key) {
+      case 'sees':
+        return sees;
+      case 'owns':
+        return owns;
+      case 'status':
+        return status;
       default:
         return key;
     }
@@ -163,7 +193,6 @@ class MichaelFriendsTestMessages implements i69n.I69nMessageBundle {
 class EvaFriendsTestMessages implements i69n.I69nMessageBundle, MichaelFriendsTestMessages {
   final FriendsTestMessages _parent;
   const EvaFriendsTestMessages(this._parent);
-  String get _i69n_implements => "MichaelFriendsTestMessages";
   String get name => "Bbbbb";
   String get description => "Bb Bb Bb";
   Object operator [](String key) {
@@ -172,8 +201,6 @@ class EvaFriendsTestMessages implements i69n.I69nMessageBundle, MichaelFriendsTe
       return (this[key.substring(0, index)] as i69n.I69nMessageBundle)[key.substring(index + 1)];
     }
     switch (key) {
-      case '_i69n_implements':
-        return _i69n_implements;
       case 'name':
         return name;
       case 'description':